+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
 
-## v1.0.0 (2022-04-29)
+## v1.1.0 (2022-05-02)
+
+### Features
+
+- Show raw messages to stdout while building.
+- Add an option to use cache or not on building.
+- Generate a sample config with `dbyml --init` option.
+
+
+## [v1.0.0](https://github.com/git-ogawa/dbyml/commit/4d5bfea3d0de0cabdc5d7e50e116d67b6f9569bd) (2022-04-29)
 - v1.0.0 release